restructure URL files
