+++ src/urls/index.ts
@@ -1,7 +1,10 @@
 import { Result, err, ok } from 'neverthrow';
-import { CastId, FarcasterId } from '~/urls/identifiers';
 import { parse as rawUriParse, URIComponents } from 'uri-js';
-import { ChainId } from 'caip';
+import { FarcasterURL, UnrecognizedURL } from '~/urls/baseUrl';
+import { CastURL } from '~/urls/castUrl';
+import { ChainURL } from '~/urls/chainUrl';
+import { UserURL } from '~/urls/userUrl';
+import { URL } from '~/urls/baseUrl';
 
 export const parseUrl = (url: string, allowUnrecognized = true): Result<URL, string> => {
   // extract scheme
@@ -31,123 +34,7 @@ export const parseUrl = (url: string, allowUnrecognized = true): Result<URL, str
   }
 };
 
-export abstract class URL {
-  public abstract readonly scheme: string;
-
-  public abstract toString(): string;
-}
-
-export class UnrecognizedURL extends URL {
-  public constructor(public readonly scheme: string, public readonly fullURL: string) {
-    super();
-  }
-
-  public toString() {
-    return this.fullURL;
-  }
-}
-
-export abstract class FarcasterURL extends URL {
-  public static readonly SCHEME = 'farcaster';
-  public readonly scheme = FarcasterURL.SCHEME;
-}
-
-export class UserURL extends FarcasterURL {
-  public readonly farcasterId: FarcasterId;
-
-  public static parse(url: string): Result<UserURL, string> {
-    const schemePrefix = this.SCHEME + '://';
-
-    if (!url.startsWith(schemePrefix)) {
-      return err(`URL missing 'farcaster' scheme`);
-    }
-
-    const remainder = url.substring(url.indexOf(schemePrefix) + schemePrefix.length);
-
-    const maybeFarcasterIdParams = FarcasterId.parse(remainder);
-    return maybeFarcasterIdParams.map((userIdParams) => {
-      const userId = new FarcasterId(userIdParams);
-      return new UserURL(userId);
-    });
-  }
-
-  public constructor(farcasterId: FarcasterId) {
-    super();
-    this.farcasterId = farcasterId;
-  }
-
-  public toString(): string {
-    return FarcasterURL.SCHEME + '://' + this.farcasterId.toString();
-  }
-}
-
-export class CastURL extends FarcasterURL {
-  public readonly castId: CastId;
-
-  public static parse(url: string): Result<CastURL, string> {
-    const schemePrefix = this.SCHEME + '://';
-
-    if (!url.startsWith(schemePrefix)) {
-      return err(`URL missing 'farcaster' scheme`);
-    }
-
-    const remainder = url.substring(url.indexOf(schemePrefix) + schemePrefix.length);
-
-    const maybeCastIdParams = CastId.parse(remainder);
-    return maybeCastIdParams.map((castIdParams) => {
-      const castId = new CastId(castIdParams);
-      return new CastURL(castId);
-    });
-  }
-
-  public constructor(castId: CastId) {
-    super();
-    this.castId = castId;
-  }
-
-  public toString(): string {
-    return FarcasterURL.SCHEME + '://' + this.castId.toString();
-  }
-}
-
-export abstract class BaseChainURL extends URL {
-  public static readonly SCHEME = 'chain';
-  public readonly scheme = ChainURL.SCHEME;
-}
-
-export class ChainURL extends BaseChainURL {
-  public readonly chainId: ChainId;
-
-  public static parse(url: string): Result<ChainURL, string> {
-    const schemePrefix = this.SCHEME + '://';
-
-    if (!url.startsWith(schemePrefix)) {
-      return err(`URL missing 'chain' scheme`);
-    }
-
-    const remainder = url.substring(url.indexOf(schemePrefix) + schemePrefix.length);
-
-    try {
-      const chainIdParams = ChainId.parse(remainder);
-      const chainId = new ChainId(chainIdParams);
-
-      // check for extra invalid data before or after the chain ID
-      const referenceRegex = new RegExp('^' + ChainId.spec.parameters.values[1].regex + '$');
-      if (!referenceRegex.test(chainId.reference)) {
-        return err(`ChainURL.parse: invalid extra data after ChainId: '${remainder}`);
-      }
-      return ok(new ChainURL(chainId));
-    } catch (e) {
-      return err(`ChainURL.parse: unable to parse '${url}`);
-    }
-  }
-
-  public constructor(chainId: ChainId) {
-    super();
-    this.chainId = chainId;
-  }
-
-  public toString(): string {
-    return BaseChainURL.SCHEME + '://' + this.chainId.toString();
-  }
-}
+export { URL } from '~/urls/baseUrl';
+export { CastURL } from '~/urls/castUrl';
+export { ChainURL } from '~/urls/chainUrl';
+export { UserURL } from '~/urls/userUrl';
