+++ src/urls/test/utils.ts
@@ -1,3 +1,4 @@
-export const expectInstanceOf = <T>(obj: any, Klass: new (...args: any[]) => T): asserts obj is T => {
+// eslint-disable-next-line prefer-arrow-functions/prefer-arrow-functions
+export function expectInstanceOf<T>(obj: any, Klass: new (...args: any[]) => T): asserts obj is T {
   expect(obj).toBeInstanceOf(Klass);
-};
+}
